+++ src/include/CameraModel.hpp
@@ -10,23 +10,23 @@ namespace targetfinder {
     class CameraModel {
         friend class TargetFinder;
     public:
-        static constexpr float FOCAL_LENGTH = 3.6;
-        static constexpr float SENSOR_WIDTH = 3.67;
-        static constexpr float SENSOR_HEIGHT = 2.74;
-        static constexpr float MAGNIFICATION_FACTOR = 1.0;
-        static constexpr float TARGET_WIDTH = 594;
-        static constexpr float TARGET_HEIGHT = 420;
+        static constexpr double FOCAL_LENGTH = 3.6;
+        static constexpr double SENSOR_WIDTH = 3.67;
+        static constexpr double SENSOR_HEIGHT = 2.74;
+        static constexpr double MAGNIFICATION_FACTOR = 1.0;
+        static constexpr double TARGET_WIDTH = 594;
+        static constexpr double TARGET_HEIGHT = 420;
 
-        float distance(int calc_width, int calc_height);
-        CameraModel(int image_width, int image_height, float target_width=TARGET_WIDTH,
-                    float target_height=TARGET_HEIGHT, float focal_length=FOCAL_LENGTH,
-                    float magnification_factor=MAGNIFICATION_FACTOR, float sensor_width=SENSOR_WIDTH,
-                    float sensor_height=SENSOR_HEIGHT);
+        double distance(int calc_width, int calc_height);
+        CameraModel(int image_width, int image_height, double target_width=TARGET_WIDTH,
+                    double target_height=TARGET_HEIGHT, double focal_length=FOCAL_LENGTH,
+                    double magnification_factor=MAGNIFICATION_FACTOR, double sensor_width=SENSOR_WIDTH,
+                    double sensor_height=SENSOR_HEIGHT);
     protected:
         int image_width, image_height;
-        float target_width = TARGET_WIDTH, target_height = TARGET_HEIGHT;
-        float focal_length = FOCAL_LENGTH, magnification_factor = MAGNIFICATION_FACTOR;
-        float sensor_width = SENSOR_WIDTH, sensor_height = SENSOR_HEIGHT;
+        double target_width = TARGET_WIDTH, target_height = TARGET_HEIGHT;
+        double focal_length = FOCAL_LENGTH, magnification_factor = MAGNIFICATION_FACTOR;
+        double sensor_width = SENSOR_WIDTH, sensor_height = SENSOR_HEIGHT;
     };
 
 }
